src/test/java/seedu/address/model/event/EventNameTest.java
package seedu.address.model.event;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

import org.junit.Test;

import seedu.address.testutil.Assert;

public class EventNameTest {

    @Test
    public void constructor_null_throwsNullPointerException() {
        Assert.assertThrows(NullPointerException.class, () -> new EventName(null));
    }

    @Test
    public void constructor_invalidEventName_throwsIllegalArgumentException() {
        String invalidName = "";
        Assert.assertThrows(IllegalArgumentException.class, () -> new EventName(invalidName));
    }

    @Test
    public void isValidName() {
        // null name
        Assert.assertThrows(NullPointerException.class, () -> EventName.isValidName(null));

        // invalid name
        assertFalse(EventName.isValidName("")); // empty string
        assertFalse(EventName.isValidName(" ")); // spaces only

        // valid name
        assertTrue(EventName.isValidName("31st birthday")); // alphanumeric characters
        assertTrue(EventName.isValidName("John's Birthday!!")); // with special characters
        assertTrue(EventName.isValidName("David Roger Jackson Ray Jr 2nd's Birthday")); // long names
    }
}
